DTFS2-4615 - dashboard filters - split up, renaming, DRY, render facility filters. Allow filters to have different field name and value
