+++ portal/server/constants/facility-has-been-issued.js
@@ -1,0 +1,6 @@
+const FACILITY_HAS_BEEN_ISSUED = {
+  ISSUED: true,
+  UNISSUED: false,
+};
+
+module.exports = FACILITY_HAS_BEEN_ISSUED;
