--- portal/server/constants/facility-stage.js
@@ -1,14 +1,0 @@
-const FACILITY_STATUS = {
-  unissued_conditional: 'Unissued/Conditional',
-  issued_unconditional: 'Issued/Unconditional',
-/*
-  unissued: 'Unissued',
-  issued: 'Issued',
-  conditional: 'Conditional',
-  unconditional: 'Unconditional',
-  incomplete: 'Incomplete',
-  submitted: 'Submitted',
-*/
-};
-
-module.exports = FACILITY_STATUS;
