+++ portal/server/constants/facility-type.js
@@ -1,0 +1,8 @@
+const FACILITY_TYPE = {
+  BOND: 'Bond',
+  LOAN: 'Loan',
+  CASH: 'Cash',
+  CONTINGENT: 'Contingent',
+};
+
+module.exports = FACILITY_TYPE;
