+++ portal/server/constants/field-names.js
@@ -4,6 +4,12 @@ const DEAL = {
   STATUS: 'status',
 };
 
+const FACILITY = {
+  TYPE: 'type',
+  HAS_BEEN_ISSUED: 'hasBeenIssued',
+};
+
 module.exports = {
   DEAL,
+  FACILITY,
 };
